Polish VS4Mac editor document manager.
